add a deno script to clear tx pool
